remove unused code + move code
